ref: Split Hledger.Data.AccountBalance and clean up documentation.
This keeps Hledger.Data.AccountBalance and Hledger.Data.AccountBalances
separate.
