+++ hledger-lib/Hledger/Data.hs
@@ -11,6 +11,7 @@ functionality.  This package re-exports all the Hledger.Data.* modules
 module Hledger.Data (
                module Hledger.Data.Account,
                module Hledger.Data.AccountBalance,
+               module Hledger.Data.AccountBalances,
                module Hledger.Data.AccountName,
                module Hledger.Data.Amount,
                module Hledger.Data.Balancing,
@@ -38,6 +39,7 @@ where
 import Test.Tasty (testGroup)
 import Hledger.Data.Account
 import Hledger.Data.AccountBalance
+import Hledger.Data.AccountBalances
 import Hledger.Data.AccountName
 import Hledger.Data.Amount
 import Hledger.Data.Balancing
